requirements.txt
pandas
numpy 
ensure
PyYAML
requests
pandas 
mlflow
notebook
numpy
scikit-learn
matplotlib
pyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors
dotmap